backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
yfinance==0.2.36
ta-lib==0.4.28
alpaca-trade-api==3.2.0
joblib==1.3.2
python-multipart==0.0.9
openai==1.12.0
gspread==5.12.4
oauth2client==4.1.3